Add files via upload
This is the merged data consisting Cardekho_Extract and autoscout24-Germany-dataset with 5 columns; mileage, fuel, transmission, price, and hp. Didn't assume numbers for sum nan values, maybe when the guys doing the research method look into the trend of the data we can then decide what to do with it.   Dayo